requirements.txt
discord.py==2.1.0
PyNaCl==1.5.0
psutil==5.9.4
youtube-dl==2021.12.17
jishaku==2.5.1